Delay execution of ActiveRecord::Base-related processing
`ActiveRecord::Bitemporal` uses `ActiveSupport.on_load(:active_record) do ... end` to defer processing related to `ActiveRecord::Base`.

ref: [source](https://github.com/kufu/activerecord-bitemporal/blob/e734ab88b2c6706e1d75381dab9b03b553131184/lib/activerecord-bitemporal.rb#L164-L185)

However, there were multiple references to `ActiveRecord::Base` outside the `ActiveSupport.on_load` block, causing the deferred processing to not work as expected. To properly delay execution, the necessary files were extracted, and dependencies on `ActiveRecord::Base` are now required within the `ActiveSupport.on_load` block.
