+++ lib/activerecord-bitemporal/bitemporal.rb
@@ -1,20 +1,7 @@
 # frozen_string_literal: true
-
+require "activerecord-bitemporal/bitemporal_checker"
 module ActiveRecord
   module Bitemporal
-    module BitemporalChecker
-      refine ::Class do
-        def bi_temporal_model?
-          include?(ActiveRecord::Bitemporal)
-        end
-      end
-
-      refine ::ActiveRecord::Relation do
-        def bi_temporal_model?
-          klass.include?(ActiveRecord::Bitemporal)
-        end
-      end
-    end
     using BitemporalChecker
 
     module Optionable
